# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        example: Invalid request body
        type: string
    type: object
...
  handlers.ListResponse:
    properties:
      data: {}
    type: object
...
  handlers.SeasonResponse:
    properties:
      countingGames:
        example: 5
        type: integer
      createdAt:
        example: "2024-01-01T00:00:00Z"
        type: string
      dateCreated:
        example: "2024-01-01T00:00:00Z"
        type: string
      eventCount:
        example: 0
        type: integer
      hasFinals:
        example: false
        type: boolean
      id:
        example: 1
        type: integer
      league:
        $ref: '#/definitions/models.SwaggerLeague'
      leagueID:
        example: 1
        type: integer
      name:
        example: Spring 2024
        type: string
      pointDistribution:
        $ref: '#/definitions/models.PointDistributionMap'
      updatedAt:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
...
  models.PointDistributionMap:
    additionalProperties:
      items:
        type: number
      type: array
    type: object
...
  models.SwaggerLeague:
    properties:
      createdAt:
        example: "2024-01-01T00:00:00Z"
        type: string
      dateCreated:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      location:
        example: 123 Main St
        type: string
      name:
        example: Downtown Pinball League
        type: string
      owner:
        $ref: '#/definitions/models.SwaggerUser'
      ownerID:
        example: 1
        type: integer
      updatedAt:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
  models.SwaggerUser:
    properties:
      createdAt:
        example: "2024-01-01T00:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      firstName:
        example: John
        type: string
      id:
        example: 1
        type: integer
      lastName:
        example: Doe
        type: string
      updatedAt:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A Pinball League Management System API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pinball League API
  version: "1.0"
paths:
...
  /leagues/{leagueID}/seasons:
    get:
      description: Get a list of all seasons for a specific league
      parameters:
      - description: League ID
        in: path
        name: leagueID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of seasons
          schema:
            allOf:
            - $ref: '#/definitions/handlers.ListResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handlers.SeasonResponse'
                  type: array
              type: object
        "400":
          description: Invalid league ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List seasons for a league
      tags:
      - seasons
    post:
      consumes:
      - application/json
      description: Create a new season for a specific league
      parameters:
      - description: League ID
        in: path
        name: leagueID
        required: true
        type: string
      - description: Season details
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Season created successfully
          schema:
            allOf:
            - $ref: '#/definitions/handlers.ListResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.SeasonResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new season
      tags:
      - seasons
...
  /seasons/{seasonID}:
    get:
      description: Get detailed information about a specific season
      parameters:
      - description: Season ID
        in: path
        name: seasonID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Season details
          schema:
            $ref: '#/definitions/handlers.SeasonResponse'
        "400":
          description: Invalid season ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Season not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get a season by ID
      tags:
      - seasons
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"